Create WS_main_window.java
This is a simple game of word spelling, coded by java. When executed, the game window will occur with buttons for all English characters surrounded. 
Player and AI will take turn to guess a word, each time one character. Player can select whether player or AI will take the first guess. The winner will be the side who is forced to spell the last character of the word. Only original word is considered, i.e., only "player" is a valid word in this game, "players" is not counted.